backend/requirements.txt
# backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aioredis>=2.0.1
paramiko>=3.3.1
docker>=7.0.0
PyJWT>=2.8.0  # Replaced python-jose with PyJWT
passlib>=1.7.4
bcrypt>=4.0.1
aiohttp>=3.9.0
asyncssh>=2.14.0
psutil>=5.9.6
aiodocker>=0.21.0
prometheus-client>=0.19.0
structlog>=23.2.0